nload-0.7.4

+++ src/devreader-linux-proc.cpp
@@ -108,15 +108,15 @@ void DevReaderLinuxProc::readFromDevice(DataFrame& dataFrame)
             continue;
 
         // read device data
-        unsigned int bytesIn = 0;
-        unsigned int packetsIn = 0;
-        unsigned int errorsIn = 0;
-        unsigned int dropsIn = 0;
-        unsigned int bytesOut = 0;
-        unsigned int packetsOut = 0;
-        unsigned int errorsOut = 0;
-        unsigned int dropsOut = 0;
-        unsigned int dummy = 0;
+        unsigned long long bytesIn = 0;
+        unsigned long long packetsIn = 0;
+        unsigned long long errorsIn = 0;
+        unsigned long long dropsIn = 0;
+        unsigned long long bytesOut = 0;
+        unsigned long long packetsOut = 0;
+        unsigned long long errorsOut = 0;
+        unsigned long long dropsOut = 0;
+        unsigned long long dummy = 0;
 
         istringstream sin(trim(line.substr(posEnd + 1)));
